+++ geopositionmap/tests/test_geoposition.py
@@ -1,4 +1,4 @@
-#from decimal import float
+from decimal import Decimal
 from django.test import SimpleTestCase
 from geopositionmap import LatLng
 from example.models import POI
@@ -6,26 +6,26 @@ from example.models import POI
 
 class GeopositionTestCase(SimpleTestCase):
     def test_init_with_decimals(self):
-        gp = LatLng(float('52.5'), float('-13.4'))
-        self.assertEqual(gp.lat, float('52.5'))
-        self.assertEqual(gp.lng, float('-13.4'))
+        gp = LatLng(Decimal('52.5'), Decimal('-13.4'))
+        self.assertEqual(gp.lat, Decimal('52.5'))
+        self.assertEqual(gp.lng, Decimal('-13.4'))
 
     def test_init_with_strs(self):
         gp = LatLng('52.5', '-13.4')
-        self.assertEqual(gp.lat, float('52.5'))
-        self.assertEqual(gp.lng, float('-13.4'))
+        self.assertEqual(gp.lat, Decimal('52.5'))
+        self.assertEqual(gp.lng, Decimal('-13.4'))
 
     def test_init_with_setter(self):
         gp = LatLng()
         gp.lat = 52.5
         gp.lng = -13.4
-        self.assertEqual(gp.lat, float('52.5'))
-        self.assertEqual(gp.lng, float('-13.4'))
+        self.assertEqual(gp.lat, Decimal('52.5'))
+        self.assertEqual(gp.lng, Decimal('-13.4'))
 
     def test_init_with_floats(self):
         gp = LatLng(52.5, -13.4)
-        self.assertEqual(gp.lat, float('52.5'))
-        self.assertEqual(gp.lng, float('-13.4'))
+        self.assertEqual(gp.lat, Decimal('52.5'))
+        self.assertEqual(gp.lng, Decimal('-13.4'))
 
         
     def test_repr(self):
